copied the 'data' array to a newly created Data.js file and imported it to the Article.js file
